requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
pymysql==1.1.0
cryptography==41.0.8
email-validator==2.1.0
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
python-dateutil==2.8.2 